chore: move SubscriptionsScreen and its viewModel into ui package
