+++ composeApp/src/commonMain/kotlin/com/biondic/rssreader/article/di/ArticleModule.kt
@@ -1,6 +1,6 @@
 package com.biondic.rssreader.article.di
 
-import com.biondic.rssreader.article.ArticleViewModel
+import com.biondic.rssreader.article.ui.ArticleViewModel
 import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.module
