+++ composeApp/src/commonMain/kotlin/com/biondic/rssreader/core/ui/AppContainer.kt
@@ -3,7 +3,7 @@ package com.biondic.rssreader.core.ui
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import cafe.adriel.voyager.navigator.Navigator
-import com.biondic.rssreader.subscriptions.SubscriptionsScreen
+import com.biondic.rssreader.subscriptions.ui.SubscriptionsScreen
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
